feat: acl owner

+++ app/Http/Controllers/Admin/EndpointController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreUpdateEndpointRequest;
 use App\Models\Endpoint;
 use App\Models\Site;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class EndpointController extends Controller
 {
@@ -16,6 +17,11 @@ class EndpointController extends Controller
         if (!$site) {
             return back();
         }
+        $this->authorize('owner', $site);
+        // if (Gate::allows('owner', $site)) {
+        // if (Gate::denies('owner', $site)) {
+        //     return back();
+        // }
 
         $endpoints = $site->endpoints;
 
@@ -27,12 +33,14 @@ class EndpointController extends Controller
         if (!$site = Site::find($siteId)) {
             return back();
         }
+        $this->authorize('owner', $site);
 
         return view('admin.endpoints.create', compact('site'));
     }
 
     public function store(StoreUpdateEndpointRequest $request, Site $site)
     {
+        $this->authorize('owner', $site);
         $site->endpoints()->create($request->all());
 
         return redirect()
@@ -42,11 +50,13 @@ class EndpointController extends Controller
 
     public function edit(Site $site, Endpoint $endpoint)
     {
+        $this->authorize('owner', $site);
         return view('admin.endpoints.edit', compact('site', 'endpoint'));
     }
 
     public function update(StoreUpdateEndpointRequest $request, Site $site, Endpoint $endpoint)
     {
+        $this->authorize('owner', $site);
         $endpoint->update($request->validated());
 
         return redirect()
@@ -56,6 +66,7 @@ class EndpointController extends Controller
 
     public function destroy(Site $site, Endpoint $endpoint)
     {
+        $this->authorize('owner', $site);
         $endpoint->delete();
 
         return redirect()
